Batteries
Looks like these never got the 4x energy buff. Didn't 4x the previous values, as that would have put them above human tech.

Reduced the size of the R04 from 83 to 70. If we want to keep the idea of the R03 having a better density than the R04, then there needs to be at least something that the R04 brings to the table. But at a size of 83, you could just install two R03s for 82 space and have *better* energy capacity. With this, while the R03 has more energy density, you need more space than an R04 to actually benefit from that.