Weapons
Mainly just adjusting sale and firing costs based on the other attributes.

Did adjust the reloads on the Choleric, Ballistic, and Acuit to not be so messy in the outfitter. Damage has been adjusted accordingly.

Slightly buffed the Acuit to be a bit more on par with the Astuit in terms of DPS, just with differing ranges.

Choleric and Ballistic turrets have kept their increased rate of fire for increased energy usage.

Shifted the Guile more toward energy usage while increasing the energy usage further to be more in line with something like a Heavy Laser.